requirements.txt
adlfs==2024.7.0
aider-chat==0.65.1
aiobotocore==2.15.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aioitertools==0.12.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alabaster==1.0.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
astor==0.8.1
astroid==3.3.5
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.2.0
autoflake==2.3.0
autopep8==2.3.1
Automat==22.10.0
azure-core==1.32.0
azure-datalake-store==0.0.53
azure-identity==1.19.0
azure-storage-blob==12.24.0
babel==2.16.0
backoff==2.2.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.9.0
botocore==1.35.36
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
ConfigArgParse==1.7
constantly==23.10.4
coverage==7.6.8
croniter==5.0.1
cryptography==41.0.7
cssselect2==0.7.0
dataclasses-json==0.5.9
decorator==4.4.2
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
enchant==0.0.1
esprima==4.0.1
fastapi==0.115.3
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
flake8==7.1.1
Flask==3.1.0
flyteidl==1.13.6
flytekit==1.13.13
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gcsfs==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grep-ast==0.4.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httpx==0.27.2
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib_metadata==7.2.1
importlib_resources==6.4.5
incremental==24.7.2
influxdb-client==1.47.0
iniconfig==2.0.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
javalang==0.13.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.28
jsonlines==4.0.0
jsonpatch==1.32
jsonpickle==4.0.0
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
language_tool_python==2.8.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libcst==1.5.0
litellm==1.52.16
mando==0.7.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.13.0
mashumaro==3.14
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai
packaging
pathspec==0.12.1
pep8==1.7.1
pexpect==4.9.0
pillow==10.4.0
pinecone-client==2.2.4
pip==24.0
portalocker==2.8.2
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pyparsing==3.1.2
pyright==1.1.355
pytest==8.1.1
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
radon==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3fs==2024.10.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.3
sentence-transformers==2.5.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.28
starlette==0.36.3
statsd==4.0.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
wadllib==1.3.6
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
